Commit #1 - Fixed the Bomb duplicate
There was a problem with the program where bombs would be registered
maybe 2 going up to 6 times because I allowed collisions to cut down and
be more probable. Howevever, this also threw off the process when it
generated from the map of values to the array, as the bombs were
registered more than once because of collissions. This means that there
was a faliure with the linkedList Sets, unless that's not how it works
and I need to fix the logic of my program immediately.